back/app/requirements.txt
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4
fastapi==0.111.0
langchain==0.3.0
langchain-openai==0.2.0
langgraph==0.2.22
langchain-core==0.3.1
langchain-community==0.3.0
pydantic-settings==2.5.2
docx2txt==0.8
faiss-cpu==1.8.0.post1
tenacity==8.1.0